refactor: BoardManagemetHelper의 일부 기능 global화 Related to: #65
